Add animations to instruction and save instructions across sessions

+++ common/src/main/java/com/fossil/fossil/block/custom_blocks/AnuPortal.java
@@ -43,22 +43,22 @@ public class AnuPortal extends Block {
     @Override
     public void animateTick(BlockState state, Level level, BlockPos pos, Random random) {
         if (random.nextInt(100) == 0) {
-            level.playLocalSound((double) pos.getX() + 0.5, (double) pos.getY() + 0.5, (double) pos.getZ() + 0.5, SoundEvents.PORTAL_AMBIENT, SoundSource.BLOCKS, 0.5f, random.nextFloat() * 0.4f + 0.8f, false);
+            level.playLocalSound(pos.getX() + 0.5, pos.getY() + 0.5, pos.getZ() + 0.5, SoundEvents.PORTAL_AMBIENT, SoundSource.BLOCKS, 0.5f, random.nextFloat() * 0.4f + 0.8f, false);
         }
         for (int i = 0; i < 4; ++i) {
-            double d = (double) pos.getX() + random.nextDouble();
-            double e = (double) pos.getY() + random.nextDouble();
-            double f = (double) pos.getZ() + random.nextDouble();
-            double g = ((double) random.nextFloat() - 0.5) * 0.5;
-            double h = ((double) random.nextFloat() - 0.5) * 0.5;
-            double j = ((double) random.nextFloat() - 0.5) * 0.5;
+            double d = pos.getX() + random.nextDouble();
+            double e = pos.getY() + random.nextDouble();
+            double f = pos.getZ() + random.nextDouble();
+            double g = (random.nextFloat() - 0.5) * 0.5;
+            double h = (random.nextFloat() - 0.5) * 0.5;
+            double j = (random.nextFloat() - 0.5) * 0.5;
             int k = random.nextInt(2) * 2 - 1;
             if (level.getBlockState(pos.west()).is(this) || level.getBlockState(pos.east()).is(this)) {
-                f = (double) pos.getZ() + 0.5 + 0.25 * (double) k;
-                j = random.nextFloat() * 2.0f * (float) k;
+                f = pos.getZ() + 0.5 + 0.25 * k;
+                j = random.nextFloat() * 2.0f * k;
             } else {
-                d = (double) pos.getX() + 0.5 + 0.25 * (double) k;
-                g = random.nextFloat() * 2.0f * (float) k;
+                d = pos.getX() + 0.5 + 0.25 * k;
+                g = random.nextFloat() * 2.0f * k;
             }
             level.addParticle(ParticleTypes.PORTAL, d, e, f, g, h, j);
         }
